Save, Load & Cleanup

+++ scripts/data/IngredientData.cs
@@ -6,12 +6,12 @@ public partial class IngredientData : Resource
 {
     [Export] public string ingredientName;
 
-    [Export] public double baseQuantity;
+    [Export] public float baseQuantity;
 
     [Export] public GlobalTypes.Unit unit;
 
 
-    public IngredientData(string ingredientName, double baseQuantity, GlobalTypes.Unit unit)
+    public IngredientData(string ingredientName, float baseQuantity, GlobalTypes.Unit unit)
     {
         this.ingredientName = ingredientName;
         this.baseQuantity = baseQuantity;
